# src/main/resources/META-INF/openapi.yml
...
openapi: 3.0.1
info:
  title: Some static API
  description: Our test API description
  version: "1.2"
paths:
...
  /book/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
...
components:
  schemas:
    Book:
      type: object
      properties:
        author:
          pattern: \S
          type: string
          nullable: false
        name:
          pattern: \S
          type: string
          nullable: false
        pages:
          format: int32
          minimum: 1
          type: integer
...